New organization for files.
