+++ src/main.rs
@@ -5,9 +5,11 @@ mod assembly;
 mod balanced_tree;
 mod chunks;
 mod confidence;
+mod history_tracing;
 mod matrix;
 mod pipeline;
 mod score_params;
+mod segment_groups;
 mod segments;
 mod substitution_matrix;
 mod support;
